+++ tests/bitExpert/PHPStan/Magento/Autoload/Cache/GeneratedFileCacheUnitTest.php
@@ -18,21 +18,21 @@ use org\bovigo\vfs\vfsStreamContent;
 use org\bovigo\vfs\vfsStreamDirectory;
 use PHPUnit\Framework\TestCase;
 
-class FileCacheStorageUnitTest extends TestCase
+class GeneratedFileCacheUnitTest extends TestCase
 {
     /**
      * @var vfsStreamDirectory
      */
     private $root;
     /**
-     * @var FileCacheStorage
+     * @var GeneratedFileCache
      */
-    private $storage;
+    private $cache;
 
     public function setUp(): void
     {
         $this->root = vfsStream::setup('tmp');
-        $this->storage = new FileCacheStorage($this->root->url(), 'mage249');
+        $this->cache = new GeneratedFileCache($this->root->url(), 'mage249');
     }
 
     /**
@@ -40,7 +40,7 @@ class FileCacheStorageUnitTest extends TestCase
      */
     public function nullReturnedWhenLookingUpNonExistentFileInCache(): void
     {
-        $absFilename = $this->storage->load('test.txt', '');
+        $absFilename = $this->cache->getFile('test.txt');
 
         self::assertNull($absFilename);
     }
@@ -55,7 +55,7 @@ class FileCacheStorageUnitTest extends TestCase
             $this->root
         );
 
-        $absFilename = $this->storage->load('test.txt', '');
+        $absFilename = $this->cache->getFile('test.txt');
 
         self::assertSame($absFilename, vfsStream::url('tmp/03/ef/4b6fcb2d521ef0fd442a5301e7932d16cc9f375a.php'));
     }
@@ -65,10 +65,21 @@ class FileCacheStorageUnitTest extends TestCase
      */
     public function addingFileToCacheSucceeds(): void
     {
-        $this->storage->save('test.txt', '', 'Lorem ipsum');
-        $absFilename = $this->storage->load('test.txt', '');
+        $writtenFilename = $this->cache->putFile('test.txt', 'Lorem ipsum');
+        $absFilename = $this->cache->getFile('test.txt');
 
         self::assertSame($absFilename, vfsStream::url('tmp/03/ef/4b6fcb2d521ef0fd442a5301e7932d16cc9f375a.php'));
+        self::assertSame($absFilename, $writtenFilename);
+    }
+
+    /**
+     * @test
+     */
+    public function contentsAreWrittenToTheReturnedFile(): void
+    {
+        $writtenFilename = $this->cache->putFile('test.txt', 'Lorem ipsum');
+
+        self::assertSame('Lorem ipsum', file_get_contents($writtenFilename));
     }
 
     /**
@@ -81,8 +92,31 @@ class FileCacheStorageUnitTest extends TestCase
         // simulate full disk
         vfsStream::setQuota(1);
 
-        $this->storage->save('test.txt', '', 'Lorem ipsum');
-        $this->storage->load('test.txt', '');
+        $this->cache->putFile('test.txt', 'Lorem ipsum');
+    }
+
+    /**
+     * @test
+     */
+    public function existingCacheDirectoryIsReused(): void
+    {
+        $this->cache->putFile('test.txt', 'Lorem ipsum');
+        $writtenFilename = $this->cache->putFile('test.txt', 'Dolor sit amet');
+
+        self::assertSame('Dolor sit amet', file_get_contents($writtenFilename));
+    }
+
+    /**
+     * @test
+     */
+    public function addingFileToCacheFailsWhenCacheDirectoryCannotBeCreated(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessageMatches('#^Failed to create directory#');
+
+        $this->root->chmod(0000);
+
+        $this->cache->putFile('test.txt', 'Lorem ipsum');
     }
 
     /**
@@ -95,7 +129,7 @@ class FileCacheStorageUnitTest extends TestCase
             $this->root
         );
 
-        $absFilename = $this->storage->load('test.txt', '');
+        $absFilename = $this->cache->getFile('test.txt');
 
         self::assertNull($absFilename);
     }
@@ -105,7 +139,7 @@ class FileCacheStorageUnitTest extends TestCase
      */
     public function noTemporaryFileIsLeftBehindWhenAddingFileToCache(): void
     {
-        $this->storage->save('test.txt', '', 'Lorem ipsum');
+        $this->cache->putFile('test.txt', 'Lorem ipsum');
 
         $cacheDir = $this->root->getChild('03/ef');
         self::assertInstanceOf(vfsStreamDirectory::class, $cacheDir);